Finish up token decoding (closes #37)
