+++ OidcDebugger/Views/Debug/Index.cshtml
@@ -53,17 +53,31 @@
               {{code.value}}
           </code-card>
 
-          <token-card title="Access token" class="success"
+          <code-card title="Access token" class="success"
                     v-show="accessToken.exists"
                     v-bind:value="accessToken.value">
               {{accessToken.value}}
-          </token-card>
+          </code-card>
 
-          <token-card title="ID token" class="success"
+          <decoded-token-card title="Access token (decoded)"
+                         class="success"
+                         v-show="accessToken.exists"
+                         v-bind:value="accessToken.value">
+              {{accessToken.value}}
+          </decoded-token-card>
+
+          <code-card title="ID token" class="success"
                     v-show="idToken.exists"
                     v-bind:value="idToken.value">
               {{idToken.value}}
-          </token-card>
+          </code-card>
+
+          <decoded-token-card title="ID token (decoded)"
+                         class="success"
+                         v-show="idToken.exists"
+                         v-bind:value="idToken.value">
+              {{idToken.value}}
+          </decoded-token-card>
       </div>
 
       <div class="debug__callback-header" v-if="!success">
