+++ OidcDebugger/Views/Shared/CodeCard.vue
@@ -2,7 +2,7 @@
   <div class="card card-body code-block">
     <p class="card-text white-text">
       <slot></slot>
-      <copy-button class="bottom-right" v-bind:value="value" v-show="value"></copy-button>
+      <copy-button class="bottom-right" v-bind:value="value" v-show="value" title="Copy"></copy-button>
     </p>
   </div>
 </template>
