+++ OidcDebugger/wwwroot/css/site.css
@@ -160,7 +160,6 @@
     margin: 1em 0;
     word-wrap: break-word;
     line-height: 24px;
-    padding-bottom: 37px;
 }
 
 .code-block[title] {
@@ -214,11 +213,5 @@
 .bottom-right {
     position: absolute;
     right: 3px;
-    bottom: 3px;
-}
-
-.bottom-left {
-    position: absolute;
-    left: 3px;
     bottom: 3px;
 }
